Add `database` key to requests
This is required for running queries which do not have the full
specified table name (eg: `DATABASE.SCHEMA.TABLE`).

You can avoid setting the `DATABASE` on the connection (passing in an
empty string), as long as the user that is used to connect to Snowflake
has a configured `DEFAULT_NAMESPACE`.