Extract auth and better time handling (#16)
client.rb is probably still too big, but this at least pulls out the
auth, openssl and jwt logic into it's own class

more importantly, I've added tests around the following types and
they're handled correctly. As written, the library will always return a
Time object in UTC, so times that have a timezone will have the correct
time, but will need to be localized to whatever the client runtime wants
`localtime` to be. It was that, or make the user provide a timezone, or
rely on what locatime was in the ruby vm, which seemed like it's a
recipe for accidents.
* TIME
* DATETIME
* TIMESTAMP
* TIMESTAMP_LTZ
* TIMESTAMP_NTZ
* TIMESTAMP_TZ

---------

Co-authored-by: Stoica Alexandru <alex_stoicajr@yahoo.co.uk>
Co-authored-by: Jordan Brough <jordan-brough@users.noreply.github.com>