pass an exponential backoff to retryable.retry (#30)
Pass an exponential backoff lambda for calculating the number of seconds
to sleep when we hit retryable errors. In practice, this is almost
always a 429 where we've exceeded the capacity of a warehouse
temporarily.

I couldn't come up with a great way to test this separately, so I'm
going to rely on Retryable having tested their sleep parameter well.