Use json gem instead of oj (#75)
I was reading this blog post:
https://byroot.github.io/ruby/json/2024/12/15/optimizing-ruby-json-part-1.html
where the new maintainer of the json gem says:

> I was recently made maintainer of the json gem, and aside from fixing
some old bugs, I focused quite a bit on its
> performance, so that it is now the fastest JSON parser and generator
for Ruby on most benchmarks.

Since JSON is the default Ruby gem and is also now the fastest (by his
claim, at least), it seems like it'd be nice to go with that and avoid
an extra dependency, unless we have another reason to use Oj.

---------

Co-authored-by: Stoica Alexandru <alex_stoicajr@yahoo.co.uk>